Improve dependency resolution
We now build an N-ary tree of the dependencies of the package which is
marked for installation.
This allows creating sane runtimes, as well as knowing in advance if we
actually can install a package.
This code might still have bugs, it has not yet been tested for complex
cases.