Clean up system on shutdown
This is an essential feature for the implementation of atomic upgrades:
In order to remove old resources which are currently in use, we would
have to continuously watch the cgroup the updated Limba application is
in. This is really annoying, and also not always effective, since we
will have cruft around in case the watching fails (e.g. by a sudden
poweroff of the machine).
Instead, we skip all the watching and monitoring and simply flag an
updated application as to-be-removed, and perform the real removal on
shutdown.
This should be safe and also not delay the shutdown process much.