Get rid of all unit-test hacks by creating a fake root environment
This solution unfortunately makes it necessary to run the testsuite with
root permission.
On the pro side, it removes lots of hacks which were in the Limba code
just to make testing possible.
The new testing framework should make it much easier to write tests and
the tests should be much more realistic now, than with the previous
path-hacks.
Technically, we create a new root tree, mount the host filesystem into
it using OverlayFS and keep a "shadow" copy with volatile data created
by the testsuite, where all write operations of the testsuite are
collected.
State data and /opt is not taken from the host, so Limba tests should
run unaffected by eventual other Limba packages which have been
installed on the host.

Careful! This patch has only been tested briefly so far, and maybe has
bugs which I do not see yet. The (un)mounting could also be made smarter
in future by having a separate process supervise the tests and create
the fake environment.