Implement rudimentary progress notifications
This needs some deeper thinking to be done properly, e.g. we might want
to forward download speed and possibly other progress details.
Also, to adequately determine progress, we need to know all the steps
before starting to download stuff, which is currently not how the
installer is implemented.
(But shouldn't be hard to implement that, it's on TODO for a while now)