Automatically clear TryExec key from .desktop files
The use of TryExec is not a great idea with Limba bundles, so we remove
the key to prevent sideeffects.
If there is a demand for it, we might add a smarter solution in future,
to check if files in a bundle exist, but so far I do not see a need for
this.
This change will only affect .desktop files in menus, other .desktop
files will be unaffected.