Improve key importing
First, be more conservative with the GPG config for our database of
trusted keys, which we verify against.
This makes it necessary to import keys into a temporary keyring first,
and then move it to the final trusted keyring.
This should make the keyring more secure, and adding of trusted keys
easier in future.
(Distributors could also provide a static GPG public key somewhere
instead of importing stuff from a remote source).