Write signed dsc file for Limba packages
The .dsc files can be used to upload Limba packages to a remote service,
and have them processed there.
They also allow people who did not build the package to upload Limba
packages to a repository service (via the separate signature).
Also, the repository target and uploader address can be encoded there,
which allows faster processing of the package at the webserver.