Refactor package graph
An N-ary tree isn't really what we need here anymore...

This patch makes the package-graph more powerful and a bit faster at
solving dependencies.
It also makes the embedded-packages search less useful, since it would
need to be reimplemented using a different method than simple tree-
traversal. This feature isn't used much anyway, and reducing complexity
in the install process is always nice. I can imagine adding it back if
there is enough demand for it.

The patch also enhances the cruft removal with the new dependency
solver, which will make it immediately aware of SDK packages, which now
need to be explicitly removed by the user (and do not get trashed
automatically anymore).

The changes are massive, and while the tests pass there might be more
bugs, so please test carefully.