bcli: Deal with environment variables like users expect it
Environment variables count the whole build session now, once they are
set. This is much more useful, and much more like users would expect
things to be when running the build recipe.
Due to limitations regarding environment variables (and because I didn't
want to remote-control a shell in bcli), we use a environment-file based
workaround now to make the env var stuff work.