builder: Place internal host link in /host
Using the previous approach with a hidden folder in / makes some build
scripts which (unknowingly?) use shell-expansion on "." go crazy, so we
place this in a different location.
Maybe further things from the host can be mounted there as well, and we
can establish that messing around with /host is taboo for anything
running inside the venv (maybe even hide the contents of it?).