bcli: Execute helper for installations in the build venv
After forking and pivoting into the build venv, GObject creations
randomly fail, likely because there are version mismatches between the
hosts libraries and the libraries we find in the new root environment.
So, the build executor should really not do much and rather execute
helper processes in the new environment to do tasks.
Since the only place where we do something else then executing other
tools and setting up the remaining bits was the installation and
mounting of build-deps, we now call a helper for that too.

This patch should resolve all deadlocks seen in the build process.